Calculate LL in float64.

+++ osl_dynamics/models/hmm.py
@@ -258,7 +258,11 @@ def _model_structure(config):
     )  # data not used
 
     # Log-likelihood
-    ll_layer = SeparateLogLikelihoodLayer(config.n_states, name="ll")
+    ll_layer = SeparateLogLikelihoodLayer(
+        config.n_states,
+        dtype="float64",
+        name="ll",
+    )
     ll = ll_layer([data, [mu], [D]])
 
     # Hidden state inference
